Add flag to contro tf publishing

+++ src/imageProjection.cpp
@@ -1,3 +1,4 @@
+// clang-format off
 #include "utility.hpp"
 #include "lio_sam/msg/cloud_info.hpp"
 
@@ -6,12 +7,12 @@ struct VelodynePointXYZIRT
     PCL_ADD_POINT4D
     PCL_ADD_INTENSITY;
     uint16_t ring;
-    float time;
+    double time;
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 } EIGEN_ALIGN16;
 POINT_CLOUD_REGISTER_POINT_STRUCT (VelodynePointXYZIRT,
     (float, x, x) (float, y, y) (float, z, z) (float, intensity, intensity)
-    (uint16_t, ring, ring) (float, time, time)
+    (uint16_t, ring, ring) (double, time, time)
 )
 
 struct OusterPointXYZIRT {
